refactor: category page refactoring
