+++ my-monorepo/apps/shopease/src/pages/categoryProduct/components/index.ts
@@ -1,3 +1,3 @@
 export * from "./CategoryCard";
-export * from "./FilterButton";
+export * from "./CategoryFilterButton";
 export * from "./GetCategoryTitle";
